fix #104

+++ include/zen/cursor.h
@@ -10,7 +10,7 @@
 #include "zen/scene/screen.h"
 
 struct zn_cursor {
-  int x, y;
+  double x, y;
   int hotspot_x, hotspot_y;
   bool visible;
   struct zn_screen* screen;  // nullable
@@ -23,7 +23,7 @@ struct zn_cursor {
   struct wl_listener destroy_surface_listener;
 };
 
-void zn_cursor_move_relative(struct zn_cursor* self, int dx, int dy);
+void zn_cursor_move_relative(struct zn_cursor* self, double dx, double dy);
 
 void zn_cursor_set_surface(struct zn_cursor* self, struct wlr_surface* surface,
     int hotspot_x, int hotspot_y);
